feat: extract habit form and list to widget, create sign up form
